refactor: move IT case to existed class
